Refactor preferences structure and remove unused files
- Deleted color_mode.ts, general.ts, groups.ts, rows.ts, shell_version.ts files as they are no longer needed.
- Moved relevant functions to new ui directory structure for better organization.
- Updated imports in formatters.ts to reflect new file paths.
- Consolidated color control functions into a dedicated color_page directory.
- Introduced bindings.ts to handle settings binding logic.
- Enhanced version detection logic in version.ts for GNOME Shell and Libadwaita.
