+++ src/prefs/bindings.ts
@@ -1,0 +1,30 @@
+import Gio from "gi://Gio";
+import Adw from "gi://Adw";
+import { prefsGettext } from "../utils/gettext/index.js";
+import { logErr } from "../utils/error_utils.js";
+
+/**
+ * Bind a Gio.Settings key to a property on an ActionRow-like widget.
+ *
+ * @param widget - widget that exposes the property to bind
+ * @param settings - Gio.Settings instance
+ * @param settingKey - the key to bind
+ * @param property - the widget property name
+ */
+export function bindSettingsToProperty(
+  widget: Adw.ActionRow,
+  settings: Gio.Settings,
+  settingKey: string,
+  property: string,
+) {
+  try {
+    settings!.bind(settingKey, widget, property, Gio.SettingsBindFlags.DEFAULT);
+  } catch (error: any) {
+    logErr(
+      error,
+      `${prefsGettext._("Error binding settings for")} ${widget.title}`,
+    );
+  }
+}
+
+export default { bindSettingsToProperty };
